\name{compare.methods}
\Rdversion{1.1}
\alias{compare.methods}
\title{compare methods}
\description{Plot several label placement methods on the same page.}
\usage{
compare.methods(m, ...)
}
\arguments{
  \item{m}{Vector of label placement function names.}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}