\name{direct.labels}
\Rdversion{1.1}
\alias{direct.labels}
\title{direct.labels}
\description{Calculates table of positions of each label for each panel.}
\usage{
direct.labels(x, y, subscripts, groups, debug = FALSE, method = perpendicular.lines, ...)
}
\arguments{
  \item{x}{x values of points to draw.}
  \item{y}{y values of points to draw.}
  \item{subscripts}{subscripts of groups to consider.}
  \item{groups}{vector of groups.}
  \item{debug}{logical indicating whether debug annotations should be added to the plot.}
  \item{method}{function used to choose position of labels.}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}