\name{dl.indep}
\Rdversion{1.1}
\alias{dl.indep}
\title{dl.indep}
\description{Makes a function you can use to specify the location of each group independently.}
\usage{
dl.indep(expr)
}
\arguments{
  \item{expr}{Expression that takes a subset of the d data frame, with data from only a single group, and returns the direct label position.}
}
\details{
}
\value{The constructed label position function.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}