\name{dl.panel}
\Rdversion{1.1}
\alias{dl.panel}
\title{dl.panel}
\description{Convert a normal panel into a direct label panel function.}
\usage{
dl.panel(panel)
}
\arguments{
  \item{panel}{The panel function to transform.}
}
\details{
}
\value{A new panel function that first calls panel, then calls direct.labels.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}