\name{dl}
\Rdversion{1.1}
\alias{dl}
\title{dl}
\description{Shortcut for direct label lattice plots.}
\usage{
dl(lattice.fun, data, x, groups, method = NULL, debug = FALSE, ...)
}
\arguments{
  \item{lattice.fun}{High-level lattice plot function to use.}
  \item{data}{Data frame to use.}
  \item{x}{Lattice model formula.}
  \item{groups}{To be passed to lattice as groups= argument.}
  \item{method}{Label placement method to be passed to direct.label.}
  \item{debug}{Show debugging output? to be passed to direct.label.}
  \item{\dots}{Other arguments to be passed to lattice.fun.}
}
\details{
}
\value{The lattice plot.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}