\name{dl.text}
\Rdversion{1.1}
\alias{dl.text}
\title{dl.text}
\description{
}
\usage{
dl.text(labs, group.number, col.line = NULL, col.points = NULL, col = NULL, col.symbol = NULL, type = NULL, ...)
}
\arguments{
  \item{labs}{table of labels and positions constructed by direct.labels}
  \item{group.number}{which group we are currently plotting, according to levels(labs$groups)}
  \item{col.line}{line color}
  \item{col.points}{point color}
  \item{col}{general color}
  \item{col.symbol}{symbol color}
  \item{type}{plot type}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}