\name{first.points}
\Rdversion{1.1}
\alias{first.points}
\title{first.points}
\description{Position Function for labeling first points of longitudinal data}
\usage{
first.points(d, debug)
}
\arguments{
  \item{d}{
}
  \item{debug}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}