\name{get.means}
\Rdversion{1.1}
\alias{get.means}
\title{get.means}
\description{get the mean of each cluster of points}
\usage{
get.means(d, debug)
}
\arguments{
  \item{d}{data frame with groups x y}
  \item{debug}{ignored}
}
\details{
}
\value{data frame with groups x y, 1 row for each group, with the calculated mean values}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}