\name{high.points}
\Rdversion{1.1}
\alias{high.points}
\title{high points}
\description{  dl.indep(data.frame(d[which.max(d$y),],hjust=0.5,vjust=0))}
\usage{
high.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}