\name{last.points}
\Rdversion{1.1}
\alias{last.points}
\title{last.points}
\description{Position Function for labeling last points of longitudinal data}
\usage{
last.points(d, debug)
}
\arguments{
  \item{d}{
}
  \item{debug}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}