\name{last}
\Rdversion{1.1}
\alias{last}
\title{last}
\description{Return the last data point in d as a data frame. This works for time series data when already sorted and separated into groups.}
\usage{
last(d, debug = FALSE)
}
\arguments{
  \item{d}{Data frame with columns x y}
  \item{debug}{Show debug graphics?}
}
\details{
}
\value{Data frame with 1 row}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}