\name{latticedl-package}
\Rdversion{1.1}
\alias{latticedl-package}
\docType{package}
\title{Lattice direct labels}
\description{Direct labeling functions that use the lattice package.}
\details{
\tabular{ll}{Package: \tab latticedl\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.1\cr
License: \tab GPL-3\cr
Title: \tab Lattice direct labels\cr
Description: \tab Direct labeling functions that use the lattice package.\cr
Depends: \tab plyr, lattice, grid\cr
Suggests: \tab nlme, mlmRev, ggplot2\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}