\name{left.points}
\Rdversion{1.1}
\alias{left.points}
\title{left points}
\description{  dl.indep(data.frame(d[which.min(d$x),],hjust=1,vjust=0.5))}
\usage{
left.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}