\name{low.points}
\Rdversion{1.1}
\alias{low.points}
\title{low points}
\description{  dl.indep(data.frame(d[which.min(d$y),],hjust=0.5,vjust=1))}
\usage{
low.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}