\name{most.likely}
\Rdversion{1.1}
\alias{most.likely}
\title{most.likely}
\description{Considering a density estimate of d$x, return the most likely point.}
\usage{
most.likely(d, debug)
}
\arguments{
  \item{d}{Data frame with column x.}
  \item{debug}{Show debugging graphics?}
}
\details{
}
\value{Data frame with 1 line.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}