\name{panel.dl}
\Rdversion{1.1}
\alias{panel.dl}
\title{panel dl}
\description{Panel function for adding direct labels to a plot.}
\usage{
panel.dl(method, debug, lattice.fun.name, type = NULL, ..., end = 0.03)
}
\arguments{
  \item{method}{The direct labeling method.}
  \item{debug}{Show debugging output?}
  \item{lattice.fun.name}{used for default Positioning Function dispatch.}
  \item{type}{used for default Positioning Function dispatch.}
  \item{\dots}{
}
  \item{end}{used for default Positioning Function for rugs.}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}