\name{trans.densityplot}
\Rdversion{1.1}
\alias{trans.densityplot}
\title{trans densityplot}
\description{Transformation function for 1d densityplots.}
\usage{
trans.densityplot(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}