\name{trans.qqmath}
\Rdversion{1.1}
\alias{trans.qqmath}
\title{trans qqmath}
\description{Transformation function for 1d qqmath plots.}
\usage{
trans.qqmath(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}