\name{empty.grid.2}
\Rdversion{1.1}
\alias{empty.grid.2}
\title{empty.grid.2}
\description{Use the parallel lines method in combination with the empty grid method.}
\usage{
empty.grid.2(d, debug)
}
\arguments{
  \item{d}{Data frame with columns groups x y}
  \item{debug}{
}
}
\details{
}
\value{Show debugging graphics on the plot?}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}