\name{first.points}
\Rdversion{1.1}
\alias{first.points}
\title{first points}
\description{Positioning Function for first points of longitudinal data.}
\usage{
first.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}