\name{get.means}
\Rdversion{1.1}
\alias{get.means}
\title{get means}
\description{Positioning Function for the mean of each cluster of points.}
\usage{
get.means(d, debug)
}
\arguments{
  \item{d}{
}
  \item{debug}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}