\name{last.points}
\Rdversion{1.1}
\alias{last.points}
\title{last points}
\description{Positioning Function for last points of longitudinal data.}
\usage{
last.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}