\name{latticedl-package}
\Rdversion{1.1}
\alias{latticedl-package}
\docType{package}
\title{Lattice direct labels}
\description{Direct label panel functions for the lattice package.}
\details{
\tabular{ll}{Package: \tab latticedl\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@etu.upmc.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Lattice direct labels\cr
Description: \tab Direct label panel functions for the lattice package.\cr
Depends: \tab plyr, lattice, grid\cr}
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}