\name{panel.superpose.dl}
\Rdversion{1.1}
\alias{panel.superpose.dl}
\title{panel superpose dl}
\description{Call panel.superpose and then panel.dl.}
\usage{
panel.superpose.dl(x, y = NULL, subscripts, groups, panel.groups, method = NULL, debug = FALSE, ...)
}
\arguments{
  \item{x}{vector of x values.}
  \item{y}{vector of y values.}
  \item{subscripts}{subscripts of x,y,groups.}
  \item{groups}{vector of group ids.}
  \item{panel.groups}{To be parsed for default labeling method, and passed to panel.superpose.}
  \item{method}{Positioning Function to use for label placement.}
  \item{debug}{Show debug output?}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}