\name{top.points}
\Rdversion{1.1}
\alias{top.points}
\title{top points}
\description{Positioning Function for the mode of a density estimate.}
\usage{
top.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}